fabcon26/demo-provisioning/ado/requirements.txt
# =============================================================================
# Python Dependencies — Fabric CI/CD Deployment
# =============================================================================
# Install: pip install -r ado/requirements.txt
# =============================================================================

fabric-cicd>=0.1.0
azure-identity>=1.15.0
